backend/requirements.txt
# FastAPI and core dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
redis==5.0.1
hiredis==2.3.2

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==41.0.7

# Vector Database
pinecone-client==3.0.0

# Message Queue
aio-pika==9.3.1

# AI/ML Integration
numpy==1.26.3
torch==2.1.2
sentence-transformers==2.3.1
transformers==4.37.0

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
websockets==12.0
